indexing: update patron types mapping
* Adds analyzer on name and description fields.
