+++ rero_ils/modules/patron_types/mappings/v6/patron_types/patron_type-v0.0.1.json
@@ -17,13 +17,15 @@
         },
         "name": {
           "type": "text",
-          "copy_to": "patron_type_name"
+          "copy_to": "patron_type_name",
+          "analyzer": "global_lowercase_asciifolding"
         },
         "patron_type_name": {
           "type": "keyword"
         },
         "description": {
-          "type": "text"
+          "type": "text",
+          "analyzer": "global_lowercase_asciifolding"
         },
         "organisation": {
           "properties": {
